requirements.txt
# Database
psycopg[binary]>=3.1.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client (for vLLM API)
requests>=2.31.0

# PDF text extraction (no OCR; text-layer only)
pymupdf>=1.24.0

# API/Web framework (for Phase 1 API)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Embedding and recommendation
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0

# XML security
defusedxml>=0.7.1

# Development
pytest>=7.4.0
black>=24.0.0
ruff>=0.1.0
